requirements.txt
requests
schedule
python-dotenv
pytz
email-validator
python-dotenv
